frontend: shared types
