+++ frontend/src/components/CountryList/index.tsx
@@ -1,5 +1,5 @@
 import { Box } from '@mui/material';
-import { Country } from '../../types';
+import { Country } from '../../types/shared';
 import { useEffect, useState } from 'react';
 import { getAllCountries } from '../../services/countryService';
 import CountryTable from './CountryTable';
