+++ frontend/src/components/CreateAccountPage.tsx
@@ -15,7 +15,7 @@ import {
 import { FormEvent, useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import { tryCreateAccount } from '../services/accountService';
-import { UserWithToken } from '../types';
+import { UserWithToken } from '../types/shared';
 
 const minLength = 3;
 const maxLength = 100;
