+++ frontend/src/components/GameView/HintsViews.tsx
@@ -13,7 +13,7 @@ import {
 } from '@mui/material';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import { useState } from 'react';
-import { Hint, Hints } from '../../types';
+import { Hint, Hints } from '../../types/shared';
 
 interface Props {
   hints: Hints;
